+++ core/src/com/ninetyninepercentcasino/net/BJPayout.java
@@ -1,11 +1,11 @@
 package com.ninetyninepercentcasino.net;
 
 public class BJPayout extends DTO {
-	private final double winnings;
-	public BJPayout(double winnings){
+	private final int winnings;
+	public BJPayout(int winnings){
 		this.winnings = winnings;
 	}
-	public double getWinnings(){
+	public int getWinnings(){
 		return winnings;
 	}
 }
